refactor(email): use toolargs.Uint32 for UID args, not lossy int cast
Addresses the Copilot review on #1085. The four singular-uid sites parsed
the IMAP UID as uint32(toolargs.Int(args, "uid")), which silently wraps a
negative or out-of-range value into a valid-looking large UID (e.g. -1 ->
4294967295) that the != 0 / == 0 guards do not catch — risking an operation
on the wrong message. (Pre-existing behavior carried over from the old
intArg; surfaced by the migration.)

Add a singular toolargs.Uint32 / Uint32OK getter — the single-value form of
Uint32Slice's existing element coercion (rejects nil/negative/fractional/<1/
out-of-range) — and repoint all four sites onto it, so the singular uid path
validates consistently with the uids array path. Add table-driven tests for
the new getter, including the negative-no-wrap and ceiling cases.

+++ internal/channels/email/tools.go
@@ -59,7 +59,7 @@ func (t *Tools) HandleList(ctx context.Context, args map[string]any) (string, er
 
 // HandleRead reads a single email by UID.
 func (t *Tools) HandleRead(ctx context.Context, args map[string]any) (string, error) {
-	uid := uint32(toolargs.Int(args, "uid"))
+	uid := toolargs.Uint32(args, "uid")
 	folder := toolargs.String(args, "folder")
 	account := toolargs.String(args, "account")
 
@@ -186,7 +186,7 @@ func parseMarkAction(args map[string]any) MarkAction {
 	}
 	action.UIDs = toolargs.Uint32Slice(args, "uids")
 	if len(action.UIDs) == 0 {
-		if uid := uint32(toolargs.Int(args, "uid")); uid != 0 {
+		if uid := toolargs.Uint32(args, "uid"); uid != 0 {
 			action.UIDs = []uint32{uid}
 		}
 	}
@@ -219,7 +219,7 @@ func (t *Tools) HandleSend(ctx context.Context, args map[string]any) (string, er
 // HandleReply replies to an existing message with threading headers.
 func (t *Tools) HandleReply(ctx context.Context, args map[string]any) (string, error) {
 	opts := ReplyOptions{
-		UID:      uint32(toolargs.Int(args, "uid")),
+		UID:      toolargs.Uint32(args, "uid"),
 		Folder:   toolargs.String(args, "folder"),
 		Body:     toolargs.String(args, "body"),
 		ReplyAll: toolargs.Bool(args, "reply_all"),
@@ -310,7 +310,7 @@ func (t *Tools) HandleMove(ctx context.Context, args map[string]any) (string, er
 	// Parse UIDs — accept "uids" array or singular "uid".
 	opts.UIDs = toolargs.Uint32Slice(args, "uids")
 	if len(opts.UIDs) == 0 {
-		if uid := uint32(toolargs.Int(args, "uid")); uid != 0 {
+		if uid := toolargs.Uint32(args, "uid"); uid != 0 {
 			opts.UIDs = []uint32{uid}
 		}
 	}
